add API SDK e2e tests to CI
Split the v2 e2e suite into a public tier and a funded tier.

Public tier (test:v2:public): data, errors, subaccount and multichain
specs run without secrets - an ephemeral signer is generated when
GMX_TEST_PRIVATE_KEY is not set. Wired into pr-tests.yml as a step.

Funded tier (test:v2): the three tests that submit real orders moved to
errorsLifecycle.spec.ts and subaccountLifecycle.spec.ts; the full suite
runs on pushes to release branches via sdk-e2e.yml with a funded wallet
from repo secrets. Bridge tests stay behind GMX_TEST_SEND, enabled via
a workflow_dispatch input.

Also fixes the stale /v1 suffix in sdk/.env.example.
