+++ sdk/.env.example
@@ -4,10 +4,13 @@
 #  - USDC on Base (source) for bridge-in test
 #  - ETH on Base for LayerZero nativeFee + gas
 #  - some funds on Arbitrum for bridge-out
+# Optional for the public suite (test:v2:public) — an ephemeral throwaway
+# signer is generated when unset.
 GMX_TEST_PRIVATE_KEY=0x...
 
-# API instance with multichain endpoints (e.g. local gmx-api or staging)
-GMX_TEST_API_URL=http://localhost:3004/v1
+# API instance with multichain endpoints (e.g. local gmx-api or staging).
+# No /v1 suffix — SDK request paths already include it.
+GMX_TEST_API_URL=http://localhost:3004
 
 # Settlement chain RPC (Arbitrum). Required for classic submit flows.
 GMX_TEST_RPC_URL=https://arb1.arbitrum.io/rpc
