+++ sdk/vitest.v2.public.config.ts
@@ -1,0 +1,32 @@
+import path from "path";
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { loadEnv } from "vite";
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    environment: "node",
+    testTimeout: 90_000,
+    fileParallelism: false,
+    maxConcurrency: 1,
+    include: [
+      "src/clients/v2/__tests__/data.spec.ts",
+      "src/clients/v2/__tests__/errors.spec.ts",
+      "src/clients/v2/__tests__/subaccount.spec.ts",
+      "src/clients/v2/__tests__/multichain.spec.ts",
+    ],
+    exclude: ["**/build/**", "**/node_modules/**"],
+    // eslint-disable-next-line no-restricted-globals
+    env: loadEnv("test", process.cwd(), ""),
+  },
+  resolve: {
+    alias: {
+      configs: path.resolve(__dirname, "./src/configs"),
+      utils: path.resolve(__dirname, "./src/utils"),
+      codegen: path.resolve(__dirname, "./src/codegen"),
+      abis: path.resolve(__dirname, "./src/abis"),
+      clients: path.resolve(__dirname, "./src/clients"),
+      test: path.resolve(__dirname, "./src/test"),
+    },
+  },
+});
